feat: integrate TCN with main function
